# swagger/server/pds/flaskConnexion/swagger_server/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "An API that interacts with a SQL database backend of medical provider\
    \ data system."
  version: "1.0.0"
  title: "Provider Data System"
  contact:
    name: "Scott Steinbruegge"
host: "localhost:8080"
basePath: "/PDS"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /provider:
    get:
      description: "Returns all medical providers in the system."
      operationId: "get_providers"
      parameters: []
      responses:
        200:
          description: "A list of all medical providers."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Provider"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_400"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
    put:
      description: "User can insert a new provider record."
      operationId: "create_provider"
      parameters:
      - in: "body"
        name: "provider"
        description: "Creates a new medical provider record."
        required: false
        schema:
          $ref: "#/definitions/Provider"
      responses:
        200:
          description: "Successfully created a new artist"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_400"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /provider/{npi}:
    get:
      description: "Returns medical provider of the speicifed provider NPI"
      operationId: "get_provider"
      parameters:
      - name: "npi"
        in: "path"
        description: "Provider type of providers to return"
        required: true
        type: "integer"
      responses:
        200:
          description: "Providers with the specified provider type."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Provider"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_400"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
    delete:
      description: "Delete a single provider based on NPI supplied."
      operationId: "delete_provider"
      parameters:
      - name: "npi"
        in: "path"
        description: "NPI of provider to delete"
        required: true
        type: "integer"
      responses:
        204:
          description: "Provider deleted"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_400"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
    patch:
      description: "Update a providers information based on NPI supplied"
      operationId: "update_provider"
      parameters:
      - name: "npi"
        in: "path"
        description: "NPI of provider to update"
        required: true
        type: "integer"
      - in: "body"
        name: "provider"
        description: "Data for update"
        required: false
        schema:
          $ref: "#/definitions/Provider"
      responses:
        200:
          description: "Returns the updated provider data."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Provider"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/inline_response_400"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
definitions:
  Provider:
    type: "object"
    required:
    - "npi"
    properties:
      npi:
        type: "integer"
      provider_type:
        type: "string"
      first_name:
        type: "string"
      last_name:
        type: "string"
      ssn:
        type: "integer"
    example:
      provider_type: "provider_type"
      npi: 0
      last_name: "last_name"
      first_name: "first_name"
      ssn: 6
  inline_response_400:
    properties:
      message:
        type: "string"
responses:
  400Error:
    description: "Invalid request"
    schema:
      $ref: "#/definitions/inline_response_400"